+++ src/app/(auth)/login/page.tsx
@@ -1,56 +1,46 @@
 'use client'
 
-import { useRouter, useSearchParams } from 'next/navigation'
+import { Suspense } from 'react'
+import { useRouter } from 'next/navigation'
 import { MemberLogin } from '@/components/auth/login/MemberLogin'
-import TextButton from '@/components/common/Button/TextButton'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/common/Tab/Tab'
 import { SocialLogin } from '@/components/auth/login/SocialLogin'
 import { GuestOrderCheck } from '@/components/auth/login/GuestOrderCheck'
 import { ROUTES } from '@/constants/routes'
-import { Suspense } from 'react'
 import LoadingSpinner from '@/components/common/Loading/LoadingSpinner'
+import { LoginPageParams } from '@/components/login/LoginParams'
 
 export default function LoginPage() {
-  const searchParams = useSearchParams()
-  const pathname = searchParams.get('guest')
   const router = useRouter()
+
   const handlePopularPage = () => {
     router.push(ROUTES.PURCHASE)
   }
 
   return (
-    <Suspense fallback={<LoadingSpinner />}>
-      <div>
-        <div id="section" className="flex justify-center items-center">
-          <div className="w-[438px] flex flex-col justify-center">
-            <div className="typo-h1 flex justify-center py-4">로그인</div>
-            <Tabs defaultValue="member" className="mt-8">
-              <TabsList>
-                <TabsTrigger value="member">회원</TabsTrigger>
-                <TabsTrigger value="guest">비회원 주문 조회</TabsTrigger>
-              </TabsList>
-              <TabsContent value="member">
-                <MemberLogin />
-              </TabsContent>
-              <TabsContent value="guest">
-                <GuestOrderCheck />
-              </TabsContent>
-            </Tabs>
-            <div className="my-10">
-              <SocialLogin />
-            </div>
-            {pathname == 'guestOrder' && (
-              <TextButton
-                iconPosition="right"
-                className="py-3 typo-button1 text-assistive flex"
-                onClick={handlePopularPage}
-              >
-                비회원 주문하기
-              </TextButton>
-            )}
-          </div>
+    <div id="section" className="flex justify-center items-center">
+      <div className="w-[438px] flex flex-col justify-center">
+        <div className="typo-h1 flex justify-center py-4">로그인</div>
+        <Tabs defaultValue="member" className="mt-8">
+          <TabsList>
+            <TabsTrigger value="member">회원</TabsTrigger>
+            <TabsTrigger value="guest">비회원 주문 조회</TabsTrigger>
+          </TabsList>
+          <TabsContent value="member">
+            <MemberLogin />
+          </TabsContent>
+          <TabsContent value="guest">
+            <GuestOrderCheck />
+          </TabsContent>
+        </Tabs>
+        <div className="my-10">
+          <SocialLogin />
         </div>
+
+        <Suspense fallback={<LoadingSpinner />}>
+          <LoginPageParams handlePopularPage={handlePopularPage} />
+        </Suspense>
       </div>
-    </Suspense>
+    </div>
   )
 }
